+++ src/library/DIPS.Mobile.UI/Components/Lists/CollectionView/iOS/CollectionView2Handler.cs
@@ -60,6 +60,10 @@ public class ReorderableItemsViewController(
 {
     private UIEdgeInsets? m_overridenContentInset;
 
+    private readonly Dictionary<int, Divider> m_currentDividerSetToInvisibleInSection = new();
+    private readonly Dictionary<int, UICollectionViewCell> m_currentLastCellWithCornerRadiusInSection = new();
+    private readonly Dictionary<int, UICollectionViewCell> m_currentFirstCellWithCornerRadiusInSection = new();
+    
     private void SetContentInset()
     {
         var bottomPadding = mauiCollectionView.HasAdditionalSpaceAtTheEnd
@@ -93,13 +97,18 @@ public class ReorderableItemsViewController(
         var cell = base.GetCell(collectionView, indexPath);
 
         ReordableItemsViewController.TrySetMarginOnCell(cell, mauiCollectionView);
-        ReordableItemsViewController.TrySetCornerRadiusOnCell(collectionView, indexPath, cell, mauiCollectionView);
-
-        /*if (cell.RunThroughAllChildrenUntilMatch(view => view is ContentView { CrossPlatformLayout: Divider }, out var divider))
-        {
-            
-        }*/
+        ReordableItemsViewController.TrySetCornerRadiusOnCell(collectionView, indexPath, cell, mauiCollectionView, m_currentLastCellWithCornerRadiusInSection, m_currentFirstCellWithCornerRadiusInSection);
+        ReordableItemsViewController.TrySetDividerInvisible(collectionView, indexPath, cell, mauiCollectionView, m_currentDividerSetToInvisibleInSection);
         
         return cell;
+    }
+
+    protected override void Dispose(bool disposing)
+    {
+        base.Dispose(disposing);
+        
+        m_currentDividerSetToInvisibleInSection.Clear();
+        m_currentFirstCellWithCornerRadiusInSection.Clear();
+        m_currentLastCellWithCornerRadiusInSection.Clear();
     }
 }